overheat: Monitor USB accessories for 150 seconds after insertion
Currently, the driver only mitigates overheat conditions when a PD
charger is connected, but a lot of current can still go across the port
when a USB accessory (e.g. a digital audio adapter) is connected. A
short would cause the port to heat up fairly quickly, so it is only
necessary to monitor it for the first couple of minutes.

Bug: 111480677
Test: Connected condor and observed logs
Change-Id: Ia9ba81f9dc0bd64df2a3d5636e3d35ef9e75c45d
Signed-off-by: Maggie White <maggiewhite@google.com>