smb-lib: Dont report pd current while negotiation
This CL makes current_max to reflect pd_current_max only
when the contract negotiation is done. This is done to
remove transient charging notifications. This is done
to hide psinkstndby power transitions to the user.

From the spec:

7.2.3 Sink Standby
The Sink Shall transition to Sink Standby before a
positive or negative voltage transition of V BUS
During Sink Standby the Sink Shall reduce its power draw
to pSnkStdby. This allows the Source to manage the voltage
transition as well as supply sufficient operating current
to the Sink to maintain PD operation during the transition.
The Sink Shall complete this transition to Sink Standby
within tSnkStdby after evaluating the Accept Message
from the Source. The transition when returning to Sink
operation from Sink Standby Shall be completed within
tSnkNewPower. The pSnkStdby requirement Shall only
apply if the Sink power draw is higher than this level.

Bug: 112517205
Change-Id: I8e1d2a2f0fba539c39e2fdf2608b1e3385498850
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>